SortOptions: refactored to overmind (#3183)
