+++ packages/app/src/app/pages/Dashboard/Content/Sandboxes/Filters/SortOptions/Option.tsx
@@ -1,9 +1,23 @@
 import React from 'react';
 import Check from 'react-icons/lib/md/check';
+
 import { IconContainer, OptionContainer } from './elements';
 
-export default ({ name, field, currentField, setField }) => {
+interface Props {
+  name: string;
+  field: string;
+  currentField: string;
+  setField: (field: string) => void;
+}
+
+export const Option: React.FC<Props> = ({
+  name,
+  field,
+  currentField,
+  setField,
+}) => {
   const selected = field === currentField;
+
   return (
     <OptionContainer onClick={() => setField(field)} selected={selected}>
       <IconContainer>{selected && <Check />}</IconContainer> {name}
